Move creation of Rolodex and RolodexView to the ApplicationView
