+++ src/app.js
@@ -7,30 +7,7 @@ import ContactView from 'app/views/contact_view';
 import Rolodex from 'app/collections/rolodex';
 import RolodexView from 'app/views/rolodex_view';
 
-var contactData = [
-  {
-    name: 'Freddie',
-    phoneNumber: 123456,
-    email: 'fred@msn.com'
-  },
-  {
-    name: 'Shadow',
-    phoneNumber: 5555555,
-    email: 'shadow@example.com'
-  },
-  {
-    name: 'Mochi',
-    phoneNumber: 7922222,
-    email: 'mochi@example.com'
-  }];
-
 $(document).ready(function() {
-  var contactList = new Rolodex(contactData);
-
-  var rolodexView = new RolodexView({
-    el: 'main',
-    model: contactList
-  });
 
   var application = new Application();
 
@@ -39,6 +16,6 @@ $(document).ready(function() {
     model: application
   });
 
-  // appView.render();
-  rolodexView.render();
+  appView.render();
+  // rolodexView.render();
 });
